back-end_FastApi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
fastapi-login
passlib[bcrypt]
python-jose
jinja2
rich
config
mariadb
sqlalchemy
alembic
aiofiles
babel
Flask-Script
pydantic[email]
pydantic==1.*
python-dateutil
fastapi-limiter
phpserialize
sqlalchemy_utils
PyJWT
python-decouple